+++ src/entityStore/StoreBackedItemIndex.ts
