+++ src/pagination/paginators/MessageIntervalPaginator.ts
@@ -39,8 +39,8 @@ import type { FieldToDataResolver } from '../types.normalization';
 import { resolveDotPathValue } from '../utility.normalization';
 import { lowerBound } from '../utility.search';
 import type { ItemIndexApi } from '../ItemIndex';
-import type { EntityStoreChangeBatch } from '../../messageStore/EntityStore';
-import { StoreBackedItemIndex } from '../../messageStore/StoreBackedItemIndex';
+import type { EntityStoreChangeBatch } from '../../entityStore/EntityStore';
+import { StoreBackedItemIndex } from '../../entityStore/StoreBackedItemIndex';
 import { deriveCreatedAtAroundPaginationFlags } from '../cursorDerivation';
 import { deriveIdAroundPaginationFlags } from '../cursorDerivation/idAroundPaginationFlags';
 import { deriveLinearPaginationFlags } from '../cursorDerivation/linearPaginationFlags';
